--- src/components/Card/index.js
@@ -1,11 +1,0 @@
-import React from "react";
-import "./style.css";
-
-
-function Card (props){
-
-return(<div className = "card"> {props.children}</div>)
-
-}
-
-export default Card;
