--- src/components/Forms/index.js
@@ -1,6 +1,0 @@
-export * from './User';
-export * from './Login';
-export * from './Company';
-export * from './Client';
-export * from './Project';
-export * from './Task';
