+++ src/components/Private Views/Forms/index.js
@@ -1,0 +1,4 @@
+export * from './Company';
+export * from "./Client";
+export * from './Project';
+export * from './Task';
