+++ src/components/Private Views/GlobalHeader/style.css
