+++ src/components/Private Views/Navbar/index.js
