+++ src/components/Public Views/Forms/index.js
@@ -1,0 +1,2 @@
+export * from './User';
+export * from "./Login";
