+++ src/components/Utilities/Card/index.js
@@ -1,0 +1,11 @@
+import React from "react";
+import "./style.css";
+
+
+function Card(props) {
+
+    return (<div className="card"> {props.children}</div>)
+
+}
+
+export default Card;
